Should really fix the voice clip issues this time for mac..
git-svn-id: svn://svn.code.sf.net/p/amsn/code/trunk@7389 f248a3e1-9e13-0410-b2cf-a4429de6d898

+++ amsn/utils/tcl_siren/src/encoder.h
@@ -76,16 +76,16 @@ extern int categorize_regions(int number_of_regions, int number_of_available_bit
 
 
 
-#define GUINT16_FROM_LE(val) ( (short) ( (unsigned short) IDX(val, 0) + (unsigned short) IDX(val, 1) * 256 ))
-#define GUINT32_FROM_LE(val) ( (int) (IDX(val, 0) + IDX(val, 1) * 256 + \
+#define GUINT16_FROM_LE(val) ( (unsigned short) ( IDX(val, 0) + IDX(val, 1) * 256 ))
+#define GUINT32_FROM_LE(val) ( (unsigned int) (IDX(val, 0) + IDX(val, 1) * 256 + \
         IDX(val, 2) * 65536 + IDX(val, 3) * 16777216)) 
 
 
-#define GUINT16_TO_LE(val) ( (short) (\
+#define GUINT16_TO_LE(val) ( (unsigned short) (\
         (((unsigned short)val % 256) & 0xff) << 8 | \
         ((((unsigned short)val / POW_2_8) % 256) & 0xff) ))
 
-#define GUINT32_TO_LE(val) ( (int) (\
+#define GUINT32_TO_LE(val) ( (unsigned int) (\
         ((((unsigned int) val           ) % 256)  & 0xff) << 24 | \
         ((((unsigned int) val / POW_2_8 ) % 256) & 0xff) << 16| \
         ((((unsigned int) val / POW_2_16) % 256) & 0xff) << 8 | \
